fix(dashboards): show replaced filters in the insight details popover
Move the conflict notice from a separate tile-heading badge into the
existing insight info popover's Filters section, where the applied
filters are already listed.

Generated-By: PostHog Code
Task-Id: 6e17594a-9bbb-46e4-bdf5-0473c8f6dc00
